requirements.txt
streamlit>=1.30.0
requests>=2.31.0
trafilatura>=1.6.0
PyPDF2>=3.0.0
python-docx>=1.1.0
google-generativeai>=0.3.0
edge-tts>=6.1.9
pydub>=0.25.0
python-dotenv>=1.0.0
nest_asyncio>=1.5.8
transformers>=4.40.0
torch>=2.2.0
accelerate>=0.29.0
bitsandbytes>=0.43.0
huggingface_hub>=0.23.0
openai>=1.0.0
